+++ include/igraph_paths.h
@@ -64,7 +64,8 @@ IGRAPH_EXPORT igraph_error_t igraph_get_shortest_path(const igraph_t *graph,
                                            igraph_neimode_t mode);
 
 IGRAPH_EXPORT igraph_error_t igraph_get_all_shortest_paths(const igraph_t *graph,
-                                                igraph_vector_ptr_t *res,
+                                                igraph_vector_ptr_t *vertices,
+                                                igraph_vector_ptr_t *edges,
                                                 igraph_vector_t *nrgeo,
                                                 igraph_integer_t from, const igraph_vs_t to,
                                                 igraph_neimode_t mode);
@@ -113,7 +114,8 @@ IGRAPH_EXPORT igraph_error_t igraph_get_shortest_path_dijkstra(const igraph_t *g
                                                     const igraph_vector_t *weights,
                                                     igraph_neimode_t mode);
 IGRAPH_EXPORT igraph_error_t igraph_get_all_shortest_paths_dijkstra(const igraph_t *graph,
-                                                         igraph_vector_ptr_t *res,
+                                                         igraph_vector_ptr_t *vertices,
+                                                         igraph_vector_ptr_t *edges,
                                                          igraph_vector_t *nrgeo,
                                                          igraph_integer_t from, igraph_vs_t to,
                                                          const igraph_vector_t *weights,