+++ src/paths/dijkstra.c
@@ -28,6 +28,7 @@
 #include "igraph_interface.h"
 #include "igraph_memory.h"
 #include "igraph_stack.h"
+#include "igraph_qsort.h"
 
 #include "core/indheap.h"
 #include "core/interruption.h"
@@ -118,12 +119,12 @@ int igraph_shortest_paths_dijkstra(const igraph_t *graph,
                       " of edges (%ld).", IGRAPH_EINVAL,
                       igraph_vector_size(weights), no_of_edges);
     }
+
     if (no_of_edges > 0) {
         igraph_real_t min = igraph_vector_min(weights);
         if (min < 0) {
             IGRAPH_ERRORF("Weight vector must be non-negative, got %g.", IGRAPH_EINVAL, min);
-        }
-        else if (igraph_is_nan(min)) {
+        } else if (igraph_is_nan(min)) {
             IGRAPH_ERROR("Weight vector must not contain NaN values.", IGRAPH_EINVAL);
         }
     }
@@ -294,7 +295,8 @@ int igraph_shortest_paths_dijkstra(const igraph_t *graph,
  *           not enough memory for temporary data.
  *        \cli IGRAPH_EINVVID
  *           \p from is invalid vertex id, or the length of \p to is
- *           not the same as the length of \p res.
+ *           not the same as the length of \p vertices (if not NULL) or
+ *           the length of \p edges (if not NULL)
  *        \cli IGRAPH_EINVMODE
  *           invalid mode argument.
  *        \endclist
@@ -621,9 +623,19 @@ int igraph_get_shortest_path_dijkstra(const igraph_t *graph,
  * in order. Assumes that both paths are pointers to igraph_vector_t
  * objects and that they are not empty
  */
-static int igraph_i_vector_tail_cmp(const void* path1, const void* path2) {
-    return (int) (igraph_vector_tail(*(const igraph_vector_t**)path1) -
-                  igraph_vector_tail(*(const igraph_vector_t**)path2));
+static int igraph_i_vector_tail_cmp(const void* a, const void* b) {
+    const igraph_vector_t *path_a = (const igraph_vector_t *) a;
+    const igraph_vector_t *path_b = (const igraph_vector_t *) b;
+    igraph_real_t tail_a = igraph_vector_tail(path_a);
+    igraph_real_t tail_b = igraph_vector_tail(path_b);
+
+    if (tail_a > tail_b) {
+        return 1;
+    } else if (tail_a < tail_b) {
+        return -1;
+    } else {
+        return 0;
+    }
 }
 
 /**
@@ -632,12 +644,18 @@ static int igraph_i_vector_tail_cmp(const void* path1, const void* path2) {
  * \brief All weighted shortest paths (geodesics) from a vertex.
  *
  * \param graph The graph object.
- * \param res Pointer to an initialized pointer vector, the result
- *   will be stored here in igraph_vector_t objects. Each vector
- *   object contains the vertices along a shortest path from \p from
- *   to another vertex. The vectors are ordered according to their
- *   target vertex: first the shortest paths to vertex 0, then to
- *   vertex 1, etc. No data is included for unreachable vertices.
+ * \param vertices Pointer to an initialized pointer vector or NULL.
+ *   If not NULL, then each vector object contains the vertices along a 
+ *   shortest path from \p from to another vertex. The vectors are
+ *   ordered according to their target vertex: first the shortest 
+ *   paths to vertex 0, then to vertex 1, etc. No data is included 
+ *   for unreachable vertices.
+ * \param edges Pointer to an initialized pointer vector or NULL. If 
+ *   not NULL, then each vector object contains the edges along a 
+ *   shortest path from \p from to another vertex. The vectors are 
+ *   ordered according to their target vertex: first the shortest 
+ *   paths to vertex 0, then to vertex 1, etc. No data is included for
+ *   unreachable vertices.
  * \param nrgeo Pointer to an initialized igraph_vector_t object or
  *   NULL. If not NULL the number of shortest paths from \p from are
  *   stored here for every vertex in the graph. Note that the values
@@ -670,8 +688,7 @@ static int igraph_i_vector_tail_cmp(const void* path1, const void* path2) {
  *        \cli IGRAPH_ENOMEM
  *           not enough memory for temporary data.
  *        \cli IGRAPH_EINVVID
- *           \p from is invalid vertex id, or the length of \p to is
- *           not the same as the length of \p res.
+ *           \p from is an invalid vertex id
  *        \cli IGRAPH_EINVMODE
  *           invalid mode argument.
  *        \endclist
@@ -686,7 +703,8 @@ static int igraph_i_vector_tail_cmp(const void* path1, const void* path2) {
  * \example examples/simple/igraph_get_all_shortest_paths_dijkstra.c
  */
 int igraph_get_all_shortest_paths_dijkstra(const igraph_t *graph,
-        igraph_vector_ptr_t *res,
+        igraph_vector_ptr_t *vertices,
+        igraph_vector_ptr_t *edges,
         igraph_vector_t *nrgeo,
         igraph_integer_t from, igraph_vs_t to,
         const igraph_vector_t *weights,
@@ -700,20 +718,21 @@ int igraph_get_all_shortest_paths_dijkstra(const igraph_t *graph,
     igraph_vit_t vit;
     igraph_2wheap_t Q;
     igraph_lazy_inclist_t inclist;
-    igraph_vector_t dists, order;
-    igraph_vector_ptr_t parents;
-    igraph_finally_func_t *res_item_destructor;
+    igraph_vector_t dists, order, index;
+    igraph_vector_ptr_t parents, parents_edge;
+
+    igraph_finally_func_t *old_edge_item_destructor, *old_vertices_item_destructor;
     unsigned char *is_target;
     long int i, n, to_reach;
+    igraph_bool_t free_vertices = 0;
 
     if (!weights) {
-        return igraph_get_all_shortest_paths(graph, res, nrgeo, from, to, mode);
+        return igraph_get_all_shortest_paths(graph, vertices, edges, nrgeo, from, to, mode);
     }
 
-    if (res == 0 && nrgeo == 0) {
+    if (vertices == 0 && nrgeo == 0 && edges == 0) {
         return IGRAPH_SUCCESS;
     }
-
     if (igraph_vector_size(weights) != no_of_edges) {
         IGRAPH_ERROR("Weight vector length does not match", IGRAPH_EINVAL);
     }
@@ -731,15 +750,34 @@ int igraph_get_all_shortest_paths_dijkstra(const igraph_t *graph,
      * of each vertex in the traversal */
     IGRAPH_CHECK(igraph_vector_ptr_init(&parents, no_of_nodes));
     IGRAPH_FINALLY(igraph_vector_ptr_destroy_all, &parents);
-    igraph_vector_ptr_set_item_destructor(&parents, (igraph_finally_func_t*)igraph_vector_destroy);
+    IGRAPH_VECTOR_PTR_SET_ITEM_DESTRUCTOR(&parents, igraph_vector_destroy);
+
+    /* parents_edge stores a vector for each vertex, listing the parent edges
+     * of each vertex in the traversal */
+    IGRAPH_CHECK(igraph_vector_ptr_init(&parents_edge, no_of_nodes));
+    IGRAPH_FINALLY(igraph_vector_ptr_destroy_all, &parents_edge);
+    IGRAPH_VECTOR_PTR_SET_ITEM_DESTRUCTOR(&parents_edge, igraph_vector_destroy);
+
     for (i = 0; i < no_of_nodes; i++) {
-        igraph_vector_t* parent_vec;
-        parent_vec = IGRAPH_CALLOC(1, igraph_vector_t);
+        igraph_vector_t *parent_vec, *parent_edge_vec;
+
+        parent_vec = igraph_Calloc(1, igraph_vector_t);
         if (parent_vec == 0) {
             IGRAPH_ERROR("cannot run igraph_get_all_shortest_paths", IGRAPH_ENOMEM);
         }
+        IGRAPH_FINALLY(igraph_free, parent_vec);
         IGRAPH_CHECK(igraph_vector_init(parent_vec, 0));
         VECTOR(parents)[i] = parent_vec;
+        IGRAPH_FINALLY_CLEAN(1);
+
+        parent_edge_vec = igraph_Calloc(1, igraph_vector_t);
+        if (parent_edge_vec == 0) {
+            IGRAPH_ERROR("cannot run igraph_get_all_shortest_paths", IGRAPH_ENOMEM);
+        }
+        IGRAPH_FINALLY(igraph_free, parent_edge_vec);
+        IGRAPH_CHECK(igraph_vector_init(parent_edge_vec, 0));
+        VECTOR(parents_edge)[i] = parent_edge_vec;
+        IGRAPH_FINALLY_CLEAN(1);
     }
 
     /* distance of each vertex from the root */
@@ -810,13 +848,17 @@ int igraph_get_all_shortest_paths_dijkstra(const igraph_t *graph,
             long int tto = IGRAPH_OTHER(graph, edge, minnei);
             igraph_real_t altdist = mindist + VECTOR(*weights)[edge];
             igraph_real_t curdist = VECTOR(dists)[tto];
-            igraph_vector_t *parent_vec;
+            igraph_vector_t *parent_vec, *parent_edge_vec;
 
             if (curdist < 0) {
                 /* This is the first non-infinite distance */
                 VECTOR(dists)[tto] = altdist;
+
                 parent_vec = (igraph_vector_t*)VECTOR(parents)[tto];
                 IGRAPH_CHECK(igraph_vector_push_back(parent_vec, minnei));
+                parent_edge_vec = (igraph_vector_t*)VECTOR(parents_edge)[tto];
+                IGRAPH_CHECK(igraph_vector_push_back(parent_edge_vec, edge));
+
                 IGRAPH_CHECK(igraph_2wheap_push_with_index(&Q, tto, -altdist));
             } else if (altdist == curdist && VECTOR(*weights)[edge] > 0) {
                 /* This is an alternative path with exactly the same length.
@@ -826,12 +868,19 @@ int igraph_get_all_shortest_paths_dijkstra(const igraph_t *graph,
                      * back-and-forth */
                 parent_vec = (igraph_vector_t*)VECTOR(parents)[tto];
                 IGRAPH_CHECK(igraph_vector_push_back(parent_vec, minnei));
+                parent_edge_vec = (igraph_vector_t*)VECTOR(parents_edge)[tto];
+                IGRAPH_CHECK(igraph_vector_push_back(parent_edge_vec, edge));
             } else if (altdist < curdist) {
                 /* This is a shorter path */
                 VECTOR(dists)[tto] = altdist;
+
                 parent_vec = (igraph_vector_t*)VECTOR(parents)[tto];
                 igraph_vector_clear(parent_vec);
                 IGRAPH_CHECK(igraph_vector_push_back(parent_vec, minnei));
+                parent_edge_vec = (igraph_vector_t*)VECTOR(parents_edge)[tto];
+                igraph_vector_clear(parent_edge_vec);
+                IGRAPH_CHECK(igraph_vector_push_back(parent_edge_vec, edge));
+
                 IGRAPH_CHECK(igraph_2wheap_modify(&Q, tto, -altdist));
             }
         }
@@ -885,8 +934,8 @@ int igraph_get_all_shortest_paths_dijkstra(const igraph_t *graph,
         }
     }
 
-    if (res) {
-        igraph_vector_t *path, *paths_index, *parent_vec;
+    if (vertices || edges) {
+        igraph_vector_t *path, *paths_index, *parent_vec, *parent_edge_vec;
         igraph_stack_t stack;
         long int j, node;
 
@@ -945,7 +994,7 @@ int igraph_get_all_shortest_paths_dijkstra(const igraph_t *graph,
         /* now, reconstruct the shortest paths from the parent list in the
          * order we've found the nodes during the traversal.
          * dists is being re-used as a vector where element i tells the
-         * index in res where the shortest paths leading to vertex i
+         * index in vertices where the shortest paths leading to vertex i
          * start, plus one (so that zero means that there are no paths
          * for a given vertex).
          */
@@ -953,28 +1002,64 @@ int igraph_get_all_shortest_paths_dijkstra(const igraph_t *graph,
         n = igraph_vector_size(&order);
         igraph_vector_null(paths_index);
 
-        /* clear the paths vector */
-        igraph_vector_ptr_clear(res);
-        res_item_destructor = igraph_vector_ptr_get_item_destructor(res);
-        igraph_vector_ptr_set_item_destructor(res,
-                                              (igraph_finally_func_t*)igraph_vector_destroy);
+        if (edges) {
+            igraph_vector_ptr_clear(edges);
+            old_edge_item_destructor = igraph_vector_ptr_get_item_destructor(edges);
+            igraph_vector_ptr_set_item_destructor(edges,
+                                                (igraph_finally_func_t*)igraph_vector_destroy);
+        }
+
+        if (vertices) {
+            igraph_vector_ptr_clear(vertices);
+        } else {
+            /* If the 'vertices' vector doesn't exist, then create one, in order
+             * for the algorithm to work. */
+            vertices = igraph_Calloc(1, igraph_vector_ptr_t);
+            if (vertices == 0) {
+                IGRAPH_ERROR("cannot run igraph_get_all_shortest_paths", IGRAPH_ENOMEM);
+            }
+            IGRAPH_FINALLY(igraph_free, vertices);
+            IGRAPH_CHECK(igraph_vector_ptr_init(vertices, 0));
+            IGRAPH_FINALLY(igraph_vector_ptr_destroy_all, vertices);
+            free_vertices = 1;
+
+            /* this is correct; needed to free everyhing at the end */
+            igraph_vector_ptr_set_item_destructor(vertices,
+                                                (igraph_finally_func_t*)igraph_vector_destroy);
+        }
+
+        old_vertices_item_destructor = igraph_vector_ptr_get_item_destructor(vertices);
+        igraph_vector_ptr_set_item_destructor(vertices, (igraph_finally_func_t*)igraph_vector_destroy);
 
         /* by definition, the shortest path leading to the starting vertex
          * consists of the vertex itself only */
         path = IGRAPH_CALLOC(1, igraph_vector_t);
-        if (path == 0)
+        if (path == 0) {
             IGRAPH_ERROR("cannot run igraph_get_all_shortest_paths_dijkstra",
-                         IGRAPH_ENOMEM);
+                        IGRAPH_ENOMEM);
+        }
         IGRAPH_FINALLY(igraph_free, path);
-        IGRAPH_CHECK(igraph_vector_init(path, 1));
-        IGRAPH_CHECK(igraph_vector_ptr_push_back(res, path));
-        IGRAPH_FINALLY_CLEAN(1);  /* ownership of path passed to res */
+        IGRAPH_VECTOR_INIT_FINALLY(path, 1);
         VECTOR(*path)[0] = from;
+        IGRAPH_CHECK(igraph_vector_ptr_push_back(vertices, path));
+        IGRAPH_FINALLY_CLEAN(2);  /* ownership of path passed to vertices */
+        
+        if (edges) {
+            path = igraph_Calloc(1, igraph_vector_t);
+            if (path == 0)
+                IGRAPH_ERROR("cannot run igraph_get_all_shortest_paths_dijkstra",
+                            IGRAPH_ENOMEM);
+            /* the shortest path from the source to itself is empty */
+            IGRAPH_FINALLY(igraph_free, path);
+            IGRAPH_VECTOR_INIT_FINALLY(path, 0);
+            IGRAPH_CHECK(igraph_vector_ptr_push_back(edges, path));
+            IGRAPH_FINALLY_CLEAN(2);  /* ownership of path passed to edges */
+        }
         VECTOR(*paths_index)[(long int)from] = 1;
 
         for (i = 1; i < n; i++) {
             long int m, path_count;
-            igraph_vector_t *parent_path;
+            igraph_vector_t *parent_path, *parent_path_edge;
 
             node = (long int) VECTOR(order)[i];
 
@@ -989,12 +1074,12 @@ int igraph_get_all_shortest_paths_dijkstra(const igraph_t *graph,
 
             /* we are calculating the shortest paths of node now. */
             /* first, we update the paths_index */
-            path_count = igraph_vector_ptr_size(res);
+            path_count = igraph_vector_ptr_size(vertices);
             VECTOR(*paths_index)[node] = path_count + 1;
-            /* res_end = (igraph_vector_t*)&(VECTOR(*res)[path_count]); */
 
             /* now, take the parent vertices */
             parent_vec = (igraph_vector_t*)VECTOR(parents)[node];
+            parent_edge_vec = (igraph_vector_t*)VECTOR(parents_edge)[node];
             m = igraph_vector_size(parent_vec);
 
             /*
@@ -1007,14 +1092,15 @@ int igraph_get_all_shortest_paths_dijkstra(const igraph_t *graph,
                 /* for each parent, copy the shortest paths leading to that parent
                  * and add the current vertex in the end */
                 long int parent_node = (long int) VECTOR(*parent_vec)[j];
+                long int parent_edge = (long int) VECTOR(*parent_edge_vec)[j];
                 long int parent_path_idx = (long int) VECTOR(*paths_index)[parent_node] - 1;
                 /*
                 printf("  Considering parent: %ld\n", parent_node);
-                printf("  Paths to parent start at index %ld in res\n", parent_path_idx);
+                printf("  Paths to parent start at index %ld in vertices\n", parent_path_idx);
                 */
                 IGRAPH_ASSERT(parent_path_idx >= 0);
                 for (; parent_path_idx < path_count; parent_path_idx++) {
-                    parent_path = (igraph_vector_t*)VECTOR(*res)[parent_path_idx];
+                    parent_path = (igraph_vector_t*)VECTOR(*vertices)[parent_path_idx];
                     if (igraph_vector_tail(parent_path) != parent_node) {
                         break;
                     }
@@ -1025,44 +1111,109 @@ int igraph_get_all_shortest_paths_dijkstra(const igraph_t *graph,
                                      IGRAPH_ENOMEM);
                     IGRAPH_FINALLY(igraph_free, path);
                     IGRAPH_CHECK(igraph_vector_copy(path, parent_path));
-                    IGRAPH_CHECK(igraph_vector_ptr_push_back(res, path));
-                    IGRAPH_FINALLY_CLEAN(1);  /* ownership of path passed to res */
+                    IGRAPH_FINALLY(igraph_vector_destroy, path);
+                    IGRAPH_CHECK(igraph_vector_ptr_push_back(vertices, path));
+                    IGRAPH_FINALLY_CLEAN(2);  /* ownership of path passed to vertices */
                     IGRAPH_CHECK(igraph_vector_push_back(path, node));
+
+                    if (edges) {
+                        path = igraph_Calloc(1, igraph_vector_t);
+                        if (path == 0)
+                            IGRAPH_ERROR("cannot run igraph_get_all_shortest_paths_dijkstra",
+                                     IGRAPH_ENOMEM);
+                        IGRAPH_FINALLY(igraph_free, path);
+                        if (parent_node != from) {
+                            parent_path_edge = (igraph_vector_t*)VECTOR(*edges)[parent_path_idx];
+                            IGRAPH_CHECK(igraph_vector_copy(path, parent_path_edge));
+                            IGRAPH_FINALLY(igraph_vector_destroy, path);
+                        } else {
+                            IGRAPH_VECTOR_INIT_FINALLY(path, 0);
+                        }
+                        IGRAPH_CHECK(igraph_vector_push_back(path, parent_edge));
+                        IGRAPH_CHECK(igraph_vector_ptr_push_back(edges, path));
+                        IGRAPH_FINALLY_CLEAN(2);  /* ownership of path passed to edges */
+                    }
                 }
             }
         }
 
-        /* remove the path vector's original item destructor */
-        igraph_vector_ptr_set_item_destructor(res, res_item_destructor);
-
         /* free those paths from the result vector which we won't need */
-        n = igraph_vector_ptr_size(res);
+        n = igraph_vector_ptr_size(vertices);
         j = 0;
         for (i = 0; i < n; i++) {
             igraph_real_t tmp;
-            path = (igraph_vector_t*)VECTOR(*res)[i];
+            path = (igraph_vector_t*)VECTOR(*vertices)[i];
             tmp = igraph_vector_tail(path);
             if (is_target[(long int)tmp] == 1) {
                 /* we need this path, keep it */
-                VECTOR(*res)[j] = path;
+                VECTOR(*vertices)[j] = path;
+                if (edges) {
+                    path = (igraph_vector_t*)VECTOR(*edges)[i];
+                    VECTOR(*edges)[j] = path;
+                }
+                /* make sure we don't leave any pointers in the vector that are
+                 * not valid any more because they have been moved elsewhere */
+                if (i > j) {
+                    VECTOR(*vertices)[i] = 0;
+                    if (edges) {
+                        VECTOR(*edges)[i] = 0;
+                    }
+                }
                 j++;
             } else {
                 /* we don't need this path, free it */
-                igraph_vector_destroy(path); free(path);
-            }
+                igraph_vector_destroy(path);
+                igraph_free(path);
+                VECTOR(*vertices)[i] = 0;
+                if (edges) {
+                    path = (igraph_vector_t*)VECTOR(*edges)[i];
+                    igraph_vector_destroy(path); free(path);
+                    VECTOR(*edges)[i] = 0;
+                }
+            }   
         }
-        IGRAPH_CHECK(igraph_vector_ptr_resize(res, j));
-
+        IGRAPH_CHECK(igraph_vector_ptr_resize(vertices, j));
+        if (edges) {
+            IGRAPH_CHECK(igraph_vector_ptr_resize(edges, j));
+        }
+        
         /* sort the paths by the target vertices */
-        igraph_vector_ptr_sort(res, igraph_i_vector_tail_cmp);
+        IGRAPH_VECTOR_INIT_FINALLY(&index, 0);
+        igraph_vector_ptr_sort_ind(vertices, &index, igraph_i_vector_tail_cmp);
+        IGRAPH_CHECK(igraph_vector_ptr_permute(vertices, &index));
+        if (edges) {
+            IGRAPH_CHECK(igraph_vector_ptr_permute(edges, &index));
+        }
+        igraph_vector_destroy(&index);
+        IGRAPH_FINALLY_CLEAN(1);
+
+        /* we can now restore the original destructors of the path and the
+         * edge-path vectors as we are not going to do thing any more that
+         * could potentially fail -- only cleanup is left */
+
+        if (vertices) {
+            igraph_vector_ptr_set_item_destructor(vertices, old_vertices_item_destructor);
+        }
+
+        if (edges) {
+            igraph_vector_ptr_set_item_destructor(edges, old_edge_item_destructor);
+        }
     }
 
     /* free the allocated memory */
+    if (free_vertices) {
+        igraph_vector_ptr_destroy_all(vertices);
+        igraph_free(vertices);
+        IGRAPH_FINALLY_CLEAN(2);
+    }
+
     igraph_vector_destroy(&order);
     IGRAPH_FREE(is_target);
     igraph_vector_destroy(&dists);
     igraph_vector_ptr_destroy_all(&parents);
-    IGRAPH_FINALLY_CLEAN(4);
+    igraph_vector_ptr_destroy_all(&parents_edge);
+    IGRAPH_FINALLY_CLEAN(5);
 
-    return 0;
+    return IGRAPH_SUCCESS;
 }
+